Add GoogleCloudReporter.
Move App Engine/Stackdriver labels from the RequestHandler to the
reporter.
The reporter merges the AsyncReporter and SyncReporter from
Google\Cloud into a single reporter with the `async` option.
